+++ chat/application-layer/index.tsx
@@ -13,6 +13,8 @@ import {
 } from '../websocket-layer';
 import { delay } from '../../util/util';
 
+const AUDIO_MESSAGE = 'Audio message';
+
 const messageTimeout = 10000;
 const fetchConversationTimeout = 15000;
 const fetchInboxTimeout = 30000;
@@ -69,16 +71,26 @@ type MessageStatus =
   | 'too long'
   | 'timeout'
 
-type ChatMessage = {
-  text: string
+type ChatBaseMessage = {
   from: string
   to: string
   fromCurrentUser: boolean
   id: string
   mamId?: string | undefined
   timestamp: Date
-  type: 'chat'
-}
+};
+
+type ChatAudioMessage = ChatBaseMessage & {
+  type: 'chat-audio'
+  audioUuid: string
+};
+
+type ChatTextMessage = ChatBaseMessage & {
+  type: 'chat-text'
+  text: string
+};
+
+type ChatMessage = ChatAudioMessage | ChatTextMessage;
 
 type TypingMessage = {
   from: string
@@ -451,7 +463,15 @@ const markDisplayed = async (message: ChatMessage) => {
 
 const sendMessage = async (
   recipientPersonUuid: string,
-  messageBody: string | null = null,
+  content: {
+    type: 'chat-text',
+    text: string,
+  } | {
+    type: 'chat-audio',
+    audioBase64: string,
+  } | {
+    type: 'typing',
+  },
   numTries: number = 3,
 ): Promise<
   | { message: Message, status: 'sent' }
@@ -467,50 +487,55 @@ const sendMessage = async (
 
   const id = getRandomString(40);
 
-  const fromJid = personUuidToJid(credentials.username);
+  const data = (() => {
+    if (content.type === 'typing') {
+      return {
+        message: {
+          '@xmlns': 'jabber:client',
+          '@type': 'typing',
+          '@from': personUuidToJid(credentials.username),
+          '@to': personUuidToJid(recipientPersonUuid),
+          '@id': id,
+        }
+      };
+    } else if (content.type === 'chat-text') {
+      return {
+        message: {
+          '@xmlns': 'jabber:client',
+          '@type': 'chat',
+          '@from': personUuidToJid(credentials.username),
+          '@to': personUuidToJid(recipientPersonUuid),
+          '@id': id,
+          body: content.text,
+        },
+      };
+    } else if (content.type === 'chat-audio') {
+      return {
+        message: {
+          '@xmlns': 'jabber:client',
+          '@type': 'chat',
+          '@from': personUuidToJid(credentials.username),
+          '@to': personUuidToJid(recipientPersonUuid),
+          '@id': id,
+          '@audioBase64': content.audioBase64,
+        },
+      };
+    } else {
+      throw new Error('Unhandled content type');
+    }
+  })();
 
-  const toJid = personUuidToJid(recipientPersonUuid);
-
-  const data = messageBody === null ? {
-    message: {
-      '@xmlns': 'jabber:client',
-      '@type': 'typing',
-      '@from': fromJid,
-      '@to': toJid,
-      '@id': id,
-    },
-  } : {
-    message: {
-      '@xmlns': 'jabber:client',
-      '@type': 'chat',
-      '@from': fromJid,
-      '@to': toJid,
-      '@id': id,
-      body: messageBody,
-    },
-  };
-
-  const message: Message = messageBody === null ? {
-    type: 'typing',
-    from: personUuidToJid(credentials.username),
-    to: personUuidToJid(recipientPersonUuid),
-    id,
-  } : {
-    type: 'chat',
-    from: personUuidToJid(credentials.username),
-    to: personUuidToJid(recipientPersonUuid),
-    id,
-    text: messageBody,
-    timestamp: new Date(),
-    fromCurrentUser: true,
-  };
-
-  const responseDetector = (doc: any): MessageStatus | null => {
+  const responseDetector = (doc: any): {
+    status: Exclude<MessageStatus, 'sent'>
+  } | {
+    status: Extract<MessageStatus, 'sent'>
+    audioUuid: string
+  } | null => {
     // Check duo_message_too_long
     try {
       const { duo_message_too_long: v } = doc;
       assert(v !== undefined);
-      return 'too long';
+      return { status: 'too long' };
     } catch { }
 
     // Check duo_message_not_unique
@@ -521,7 +546,7 @@ const sendMessage = async (
         },
       } = doc;
       assert(receivedQueryId === id);
-      return 'not unique';
+      return { status: 'not unique' };
     } catch { }
 
     // Check duo_message_blocked for rate-limited unverified basics
@@ -536,7 +561,7 @@ const sendMessage = async (
       assert(receivedQueryId === id);
       assert(reason === 'rate-limited-1day');
       assert(subreason === 'unverified-basics');
-      return 'rate-limited-1day-unverified-basics';
+      return { status: 'rate-limited-1day-unverified-basics' };
     } catch { }
 
     // Check duo_message_blocked for rate-limited unverified photos
@@ -551,7 +576,7 @@ const sendMessage = async (
       assert(receivedQueryId === id);
       assert(reason === 'rate-limited-1day');
       assert(subreason === 'unverified-photos');
-      return 'rate-limited-1day-unverified-photos';
+      return { status: 'rate-limited-1day-unverified-photos' };
     } catch { }
 
     // Check duo_message_blocked for generic rate-limited (no specific subreason)
@@ -564,7 +589,7 @@ const sendMessage = async (
       } = doc;
       assert(receivedQueryId === id);
       assert(reason === 'rate-limited-1day');
-      return 'rate-limited-1day';
+      return { status: 'rate-limited-1day' };
     } catch { }
 
     // Check duo_message_blocked for spam
@@ -577,7 +602,7 @@ const sendMessage = async (
       } = doc;
       assert(receivedQueryId === id);
       assert(reason === 'spam');
-      return 'spam';
+      return { status: 'spam' };
     } catch { }
 
     // Check duo_message_blocked for offensive
@@ -590,7 +615,7 @@ const sendMessage = async (
       } = doc;
       assert(receivedQueryId === id);
       assert(reason === 'offensive');
-      return 'offensive';
+      return { status: 'offensive' };
     } catch { }
 
     // Fallback for any duo_message_blocked case
@@ -601,7 +626,7 @@ const sendMessage = async (
         },
       } = doc;
       assert(receivedQueryId === id);
-      return 'blocked';
+      return { status: 'blocked' };
     } catch { }
 
     // Check duo_message_delivered
@@ -609,32 +634,72 @@ const sendMessage = async (
       const {
         duo_message_delivered: {
           '@id': receivedQueryId,
+          '@audioUuid': audioUuid,
         },
       } = doc;
       assert(receivedQueryId === id);
-      return 'sent';
+      return { status: 'sent', audioUuid };
     } catch { }
 
     return null;
   };
 
-  const status = await send<MessageStatus>({
+  const response = await send({
     data,
     responseDetector,
     timeoutMs: messageTimeout,
   });
 
-  if (message.type === 'typing') {
+  if (response === 'timeout') {
+    ;
+  } else if (content.type === 'typing') {
     return {
-      message,
+      message: {
+        type: 'typing',
+        from: personUuidToJid(credentials.username),
+        to: personUuidToJid(recipientPersonUuid),
+        id,
+      },
       status: 'sent', // Deliberately ignore timeouts for typing indicators
     };
-  } else if (status === 'sent') {
-    setInboxSent(recipientPersonUuid, message.text);
+  } else if (response.status === 'sent' && response.audioUuid) {
+    setInboxSent(recipientPersonUuid, AUDIO_MESSAGE);
+
     notify(`message-to-${recipientPersonUuid}`);
-    return { message, status };
-  } else if (status !== 'timeout') {
-    return { message: null, status };
+
+    return {
+      message: {
+        type: 'chat-audio',
+        from: personUuidToJid(credentials.username),
+        to: personUuidToJid(recipientPersonUuid),
+        id,
+        audioUuid: response.audioUuid,
+        timestamp: new Date(),
+        fromCurrentUser: true,
+      },
+      status: response.status,
+    };
+  } else if (response.status === 'sent') {
+    const text = content.type === 'chat-text' ? content.text : '';
+
+    setInboxSent(recipientPersonUuid, text);
+
+    notify(`message-to-${recipientPersonUuid}`);
+
+    return {
+      message: {
+        type: 'chat-text',
+        from: personUuidToJid(credentials.username),
+        to: personUuidToJid(recipientPersonUuid),
+        id,
+        text,
+        timestamp: new Date(),
+        fromCurrentUser: true,
+      },
+      status: response.status
+    };
+  } else if (response.status !== 'timeout') {
+    return { message: null, status: response.status };
   }
 
   // Deal with timeouts. To stop ourselves from sending the same message
@@ -646,7 +711,7 @@ const sendMessage = async (
     conversation === 'timeout' ||
     conversation[conversation.length - 1]?.id !== id
   ) {
-    return sendMessage(recipientPersonUuid, messageBody, numTries - 1);
+    return sendMessage(recipientPersonUuid, content, numTries - 1);
   } else {
     return { message: null, status: 'timeout' };
   }
@@ -717,67 +782,105 @@ const onReceiveMessage = (
   otherPersonUuid?: string,
   doMarkDisplayed?: boolean,
 ): (() => void) | undefined => {
-  const _onReceiveMessage = async (doc: any) => {
+  const unpackDoc = (doc: any) => {
     try {
       const {
         message: {
-          '@type': receivedType,
+          '@type': type,
           '@from': from,
           '@to': to,
           '@id': id,
-          body: bodyText,
+          '@audioUuid': audioUuid,
+          body: text,
         }
       } = doc;
 
-      assert(['chat', 'typing'].includes(receivedType));
-
-      const bareFrom = jidToBareJid(from)
-
-      if (otherPersonUuid !== undefined && otherPersonUuid !== bareFrom) {
-        return;
-      }
-
-      if (receivedType === 'typing' && callback !== undefined) {
-        const message: TypingMessage = {
-          from: from,
-          to: to,
-          id: id,
-          type: 'typing',
-        };
-
-        callback(message);
-      }
-
-      if (receivedType === 'typing') {
-        return;
-      }
-
-
-      const message: ChatMessage = {
-        text: bodyText,
-        from: from,
-        to: to,
-        id: id,
-        timestamp: new Date(),
-        fromCurrentUser: jidMatchesSignedInUser(from),
-        type: 'chat',
+      const base = {
+        from: from as string,
+        to: to as string,
+        id: id as string,
       };
 
-      await setInboxRecieved(bareFrom, bodyText);
-
-      if (otherPersonUuid === undefined) {
-        notify(`message-from-${bareFrom}`);
+      if (type === 'chat' && audioUuid) {
+        return {
+          ...base,
+          type: 'chat-audio' as 'chat-audio',
+          audioUuid: audioUuid,
+        };
       }
 
-      if (otherPersonUuid !== undefined && doMarkDisplayed !== false) {
-        await markDisplayed(message);
+      if (type === 'chat' && text){
+        return {
+          ...base,
+          type: 'chat-text' as 'chat-text',
+          text: text as string,
+        };
       }
 
-      if (callback !== undefined) {
-        callback(message);
+      if (type === 'typing') {
+        return {
+          ...base,
+          type: 'typing' as 'typing',
+        };
       }
-
     } catch { }
+
+    return null;
+  };
+
+  const _onReceiveMessage = async (doc: any) => {
+    const unpacked = unpackDoc(doc);
+
+    if (!unpacked) {
+      return;
+    }
+
+    const bareFrom = jidToBareJid(unpacked.from)
+
+    if (otherPersonUuid !== undefined && otherPersonUuid !== bareFrom) {
+      return;
+    }
+
+    if (unpacked.type === 'typing' && callback !== undefined) {
+      const message: TypingMessage = unpacked;
+
+      callback(message);
+    }
+
+    if (unpacked.type === 'typing') {
+      return;
+    }
+
+
+    const message: ChatMessage = unpacked.type === 'chat-text' ? {
+      ...unpacked,
+      type: 'chat-text',
+      timestamp: new Date(),
+      fromCurrentUser: jidMatchesSignedInUser(unpacked.from),
+    } : {
+      ...unpacked,
+      type: 'chat-audio',
+      timestamp: new Date(),
+      fromCurrentUser: jidMatchesSignedInUser(unpacked.from),
+    };
+
+    if (unpacked.type === 'chat-text') {
+      await setInboxRecieved(bareFrom, unpacked.text);
+    } else if (unpacked.type === 'chat-audio') {
+      await setInboxRecieved(bareFrom, AUDIO_MESSAGE);
+    }
+
+    if (otherPersonUuid === undefined) {
+      notify(`message-from-${bareFrom}`);
+    }
+
+    if (otherPersonUuid !== undefined && doMarkDisplayed !== false) {
+      await markDisplayed(message);
+    }
+
+    if (callback !== undefined) {
+      callback(message);
+    }
 
   };
 
@@ -829,7 +932,8 @@ const fetchConversation = async (
                 '@id': id,
                 '@from': from,
                 '@to': to,
-                'body': bodyText,
+                '@audioUuid': audioUuid,
+                'body': text,
               }
             }
           }
@@ -838,16 +942,29 @@ const fetchConversation = async (
 
       assert(receivedQueryId === queryId);
 
-      return {
-        text: bodyText,
-        from: from,
-        to: to,
-        id: id,
-        mamId: mamId ? mamId : undefined,
-        timestamp: new Date(timestamp),
-        fromCurrentUser: jidMatchesSignedInUser(from),
-        type: 'chat',
-      };
+      if (audioUuid) {
+        return {
+          type: 'chat-audio',
+          audioUuid: audioUuid,
+          from: from,
+          to: to,
+          id: id,
+          mamId: mamId ? mamId : undefined,
+          timestamp: new Date(timestamp),
+          fromCurrentUser: jidMatchesSignedInUser(from),
+        };
+      } else {
+        return {
+          type: 'chat-text',
+          text: text,
+          from: from,
+          to: to,
+          id: id,
+          mamId: mamId ? mamId : undefined,
+          timestamp: new Date(timestamp),
+          fromCurrentUser: jidMatchesSignedInUser(from),
+        };
+      }
     } catch {
       return null;
     }
@@ -947,7 +1064,7 @@ const refreshInbox = async (
               message: {
                 '@from': from,
                 '@to': to,
-                'body': bodyText,
+                'body': text,
               }
             }
           }
@@ -973,7 +1090,7 @@ const refreshInbox = async (
         name: '',
         matchPercentage: 0,
         imageUuid: null,
-        lastMessage: bodyText,
+        lastMessage: text,
         lastMessageRead: numUnread === '0',
         lastMessageTimestamp: new Date(timestamp),
         isAvailableUser: true,
