+++ util/util.tsx
@@ -152,7 +152,7 @@ const friendlyTimeAgo = (secondsAgo: number): string => {
 }
 
 const possessive = (s: string) => {
-  const possessiveMarker = String(s).endsWith('s') ? "'" : "'s";
+  const possessiveMarker = String(s).endsWith('s') ? "’" : "’s";
 
   return s + possessiveMarker;
 };
